Main#sell resource

+++ Assets/Scripts/Model/ResourceData.cs
@@ -4,5 +4,5 @@ using System;
 public class ResourceData {
 	public TileResourceEnum resourceType;
 	public string resourceName;
-	public int resourceAmount { get; set; }
+	public float resourceAmount { get; set; }
 }
